+++ SendBirdSDK.framework/Headers/SBDGroupChannelMemberListQuery.h
@@ -10,14 +10,27 @@
 
 @class SBDMember, SBDError;
 
+/**
+ *  The `SBDGroupChannelMemberListQuery` class is a query class for getting the list member in group channels. 
+ *  The instance of this class is created by [`createMemberListQuery`](../Classes/SBDGroupChannel.html#//api/name/createMemberListQuery) in `SBDGroupChannel` class.
+ */
 @interface SBDGroupChannelMemberListQuery : NSObject
+
+/**
+ *  Initialize SBDGroupChannelMemberListQuery instance.
+ *  This method is designated initializer.
+ *
+ *  @param  channelUrl      the channel url to get member list.
+ *  @return SBDGroupChannelMemberListQuery  The instance of this class.
+ */
+- (nullable instancetype)initWithChannelUrl:(nonnull NSString *)channelUrl NS_DESIGNATED_INITIALIZER;
 
 - (nullable instancetype)init NS_UNAVAILABLE;
 
 /**
  *  Sets the number of members per page.
  */
-@property (atomic, readonly) NSUInteger limit;
+@property (atomic) NSUInteger limit;
 
 /**
  *  Shows if there is a next page