+++ SendBirdSDK.framework/Headers/SendBirdSDK.h
@@ -38,6 +38,7 @@
 #import "SBDMessageListQuery.h"
 #import "SBDPreviousMessageListQuery.h"
 #import "SBDGroupChannelMemberListQuery.h"
+#import "SBDPublicGroupChannelListQuery.h"
 
 //! Project version number for SendBirdSDK.
 FOUNDATION_EXPORT double SendBirdSDKVersionNumber;